Move the shared `construct_frontend_url` and `construct_ui_canister_url` methods into a separate module under `dfx/src/util` folder.
