+++ src/dfx/src/util/mod.rs
@@ -26,6 +26,7 @@ pub mod assets;
 pub mod clap;
 pub mod currency_conversion;
 pub mod stderr_wrapper;
+pub mod url;
 
 const DECIMAL_POINT: char = '.';
